Main gained PEP 751 lockfile support (#767): `requirements_file` selects the
dependency file, `_filename` became a property, and `is_pylock` says whether it
is a lockfile. `from_source` picked the file up automatically but then treated
it as a flat requirements file, so pip's parser turned the TOML into 326
nonsense requirements.

Install through `uv pip install -r <file>` run from the file's directory, for
both formats. uv works out the format itself, which is what `build_pip_venv.sh`
now does too, and running from the file's directory is what makes relative
local paths inside a requirements.txt resolve. That replaces the hand-rolled
spec list and its local-path handling.

Read the Python version for a lockfile out of its own `requires-python`.
`uv pip compile` refuses a pylock.toml, so the probing used for a flat
requirements file cannot answer here, and the lockfile states it anyway. Read
with tomllib, falling back on 3.10 to scanning the lines above the first table
header, since `requires-python` is a top-level key and one field does not
justify a TOML dependency.

Skip the caller-satisfies check for a lockfile. It cannot be parsed as
requirements, and asking for one exact set of versions is a request for a
dedicated environment rather than for whatever happens to be installed here.

`examples/pylock` now runs under `test_unit_tesseract_from_source`, so its
`no_from_source` reason goes too.

+++ tests/runtime_tests/test_serve.py
@@ -302,7 +302,7 @@ def test_post_abstract_eval(http_client):
     response = http_client.post("/abstract_eval", json=payload)
 
     assert response.status_code == 200, response.text
-    assert response.json() == {"result": {"shape": [4], "dtype": "float64"}}
+    assert response.json() == {"result": {"shape": [4], "dtype": "float32"}}
 
 
 def test_post_abstract_eval_throws_validation_errors(http_client):
